PP-4293 Add page number, display size, from and to date type objects

+++ src/main/java/uk/gov/pay/connector/app/TransactionsPaginationServiceConfig.java
@@ -4,9 +4,9 @@ import io.dropwizard.Configuration;
 
 public class TransactionsPaginationServiceConfig extends Configuration {
 
-    private int displayPageSize;
+    private long displayPageSize;
 
-    public int getDisplayPageSize() {
+    public long getDisplayPageSize() {
         return displayPageSize;
     }
 }
